Alterando o tipo de dado do salario do funcionario e realizando relacionamento entre as entidades Proposta e Funcionario

+++ src/main/java/concessionaria/api/model/dto/FuncionarioDTO.java
@@ -27,7 +27,7 @@ public class FuncionarioDTO {
     private String compras;
     @NotNull
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal salarioBase;
+    private Double salarioBase;
     @NotNull
     private double percentualComissao;
     @Enumerated(EnumType.STRING)
